backend/requirements.txt
# web server
Flask
Flask-Cors
gunicorn

# AI and processing
nltk
openai
google-generativeai

# read .env and files
PyPDF2
python-dotenv